Don't load the media repo when configured to use an external media repo (#5754)
